backend/app/api/deps.py
from typing import Annotated

import jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jwt.exceptions import InvalidTokenError
from pydantic import ValidationError
from sqlalchemy.orm import Session

from app.core import security
from app.core.messages import (
    COULD_NOT_VALIDATE_CREDENTIALS,
    INACTIVE_USER,
    INSUFFICIENT_PRIVILEGES,
    USER_NOT_FOUND,
)
from app.core.config import settings
from app.db.session import get_db
from app.models.user import User
from app.schemas.token import TokenPayload

reusable_oauth2 = OAuth2PasswordBearer(
    tokenUrl=f"{settings.API_V1_STR}/login/access-token"
)

SessionDep = Annotated[Session, Depends(get_db)]
TokenDep = Annotated[str, Depends(reusable_oauth2)]


def get_current_user(session: SessionDep, token: TokenDep) -> User:
    try:
        payload = jwt.decode(
            token, settings.SECRET_KEY, algorithms=[security.ALGORITHM]
        )
        token_data = TokenPayload(**payload)
    except (InvalidTokenError, ValidationError):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail=COULD_NOT_VALIDATE_CREDENTIALS,
        )
    user = session.get(User, token_data.sub)
    if not user:
        raise HTTPException(status_code=404, detail=USER_NOT_FOUND)
    if not user.is_active:
        raise HTTPException(status_code=403, detail=INACTIVE_USER)
    return user


CurrentUser = Annotated[User, Depends(get_current_user)]


def get_current_active_superuser(current_user: CurrentUser) -> User:
    if not current_user.is_superuser:
        raise HTTPException(
            status_code=403, detail=INSUFFICIENT_PRIVILEGES
        )
    return current_user
